Rename classes and pathes
