+++ tests/Violation/Renderer/Html/RendererFactoryTest.php
@@ -1,0 +1,20 @@
+<?php
+namespace SensioLabs\DeprecationDetector\Tests\Violation\Renderer\Html;
+
+use PHPUnit_Framework_TestCase;
+use SensioLabs\DeprecationDetector\Violation\Renderer\Html\RendererFactory;
+use SensioLabs\DeprecationDetector\Violation\Renderer\MessageHelper\MessageHelper;
+use Symfony\Component\Filesystem\Filesystem;
+
+class RendererFactoryTest extends PHPUnit_Framework_TestCase
+{
+    public function testCreateRenderer()
+    {
+        $factory = new RendererFactory(new MessageHelper(), new Filesystem());
+
+        $this->assertInstanceOf(
+            'SensioLabs\DeprecationDetector\Violation\Renderer\Html\Renderer',
+            $factory->createHtmlOutputRenderer('./output.html')
+        );
+    }
+}
